last task MVC

+++ CodeFirstTask2/CodeFirstTask2/Controllers/StudentsController.cs
@@ -14,12 +14,22 @@ namespace CodeFirstTask2.Controllers
     {
         private SchoolContext db = new SchoolContext();
 
-        // GET: Students
+        // GET: Students/Index
         public ActionResult Index()
         {
-            var students = db.Students.Include(s => s.StudentDetails);
-            return View(students.ToList());
+            // Get list of Students with their StudentDetails
+            var students = db.Students.Include(s => s.StudentDetails).ToList();
+
+            // Create a list of StudentViewModel
+            var studentViewModels = students.Select(s => new StudentViewModel
+            {
+                Student = s,
+                StudentDetails = s.StudentDetails
+            }).ToList();
+
+            return View(studentViewModels);
         }
+
 
         // GET: Students/Details/5
         public ActionResult Details(int? id)
@@ -37,16 +47,18 @@ namespace CodeFirstTask2.Controllers
         }
 
         // GET: Students/Create
-        // GET: Students/Create
         public ActionResult Create()
         {
-            var student = new Student
+            var viewModel = new StudentViewModel
             {
+                Student = new Student(),
                 StudentDetails = new StudentDetails()
             };
 
-            return View(student);  // Passing an empty Student object with initialized StudentDetails
+            return View();
         }
+
+
 
 
         // POST: Students/Create
@@ -55,19 +67,19 @@ namespace CodeFirstTask2.Controllers
         // POST: Students/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "StudentId,FirstName,LastName,StudentDetails.Address,StudentDetails.DateOfBirth")] Student student)
+        public ActionResult Create(StudentViewModel viewModel)
         {
             if (ModelState.IsValid)
             {
-                if (student.StudentDetails != null)
-                {
-                    db.Students.Add(student);
-                    db.SaveChanges();
-                    return RedirectToAction("Index");
-                }
+                // Add the Student and StudentDetails to the database
+                viewModel.Student.StudentDetails = viewModel.StudentDetails;  // Set the relationship
+                db.Students.Add(viewModel.Student);  // Add Student (which also adds StudentDetails due to the relationship)
+                db.SaveChanges();  // Save changes to the database
+
+                return RedirectToAction("Index");
             }
 
-            return RedirectToAction("Index");
+            return View(viewModel);
         }
 
 
@@ -79,47 +91,53 @@ namespace CodeFirstTask2.Controllers
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
 
-            Student student = db.Students.Include(s => s.StudentDetails).FirstOrDefault(s => s.StudentId == id);
-
+            var student = db.Students.Include(s => s.StudentDetails).FirstOrDefault(s => s.StudentId == id);
             if (student == null)
             {
                 return HttpNotFound();
             }
 
-            return View(student);
+            var viewModel = new StudentViewModel
+            {
+                Student = student,
+                StudentDetails = student.StudentDetails
+            };
+
+            return View(viewModel);
         }
+
 
 
         // POST: Students/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to, for 
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         // POST: Students/Edit/5
+        // POST: Students/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "StudentId,FirstName,LastName,StudentDetails.Address,StudentDetails.DateOfBirth")] Student student)
+        public ActionResult Edit(StudentViewModel viewModel)
         {
             if (ModelState.IsValid)
             {
-                db.Entry(student).State = EntityState.Modified;
+                // Update Student
+                db.Entry(viewModel.Student).State = EntityState.Modified;
 
-                var studentDetails = db.StudentDetails.FirstOrDefault(sd => sd.StudentDetailsId == student.StudentDetails.StudentDetailsId);
+                // Update StudentDetails
+                var studentDetails = db.StudentDetails.FirstOrDefault(sd => sd.StudentDetailsId == viewModel.StudentDetails.StudentDetailsId);
                 if (studentDetails != null)
                 {
-                    studentDetails.Address = student.StudentDetails.Address;
-                    studentDetails.DateOfBirth = student.StudentDetails.DateOfBirth;
+                    studentDetails.Address = viewModel.StudentDetails.Address;
+                    studentDetails.DateOfBirth = viewModel.StudentDetails.DateOfBirth;
                     db.Entry(studentDetails).State = EntityState.Modified;
                 }
-                else
-                {
-                    db.StudentDetails.Add(student.StudentDetails);
-                }
 
-                db.SaveChanges();
+                db.SaveChanges();  // Save the changes to the database
                 return RedirectToAction("Index");
             }
 
-            return View(student);
+            return View(viewModel);
         }
+
 
 
         // GET: Students/Delete/5
@@ -129,32 +147,52 @@ namespace CodeFirstTask2.Controllers
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Student student = db.Students.Find(id);
+
+            // Get the Student and its StudentDetails
+            var student = db.Students.Include(s => s.StudentDetails).FirstOrDefault(s => s.StudentId == id);
             if (student == null)
             {
                 return HttpNotFound();
             }
-            return View(student);
+
+            var viewModel = new StudentViewModel
+            {
+                Student = student,
+                StudentDetails = student.StudentDetails
+            };
+
+            return View(viewModel);
         }
+
 
         // POST: Students/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Student student = db.Students.Find(id);
-            db.Students.Remove(student);
-            db.SaveChanges();
+            // Find the student by ID
+            var student = db.Students.Include(s => s.StudentDetails).FirstOrDefault(s => s.StudentId == id);
+            if (student != null)
+            {
+                // Remove the StudentDetails first if required (depending on cascade delete settings)
+                db.StudentDetails.Remove(student.StudentDetails);
+
+                // Remove the Student
+                db.Students.Remove(student);
+                db.SaveChanges();
+            }
+
             return RedirectToAction("Index");
         }
 
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
-                db.Dispose();
-            }
-            base.Dispose(disposing);
-        }
+
+        //protected override void Dispose(bool disposing)
+        //{
+        //    if (disposing)
+        //    {
+        //        db.Dispose();
+        //    }
+        //    base.Dispose(disposing);
+        //}
     }
 }
